Added CLI option 12 to update inventory

+++ src/db.rs
@@ -325,7 +325,7 @@ pub fn add_inventory_item(
 }
 
 // Function to update an inventory quantity
-pub fn update_inventory_quantity(conn: &Connection, item_id: f32, updated_quantity: f32) -> Result<()> {
+pub fn update_inventory_quantity(conn: &Connection, item_id: i32, updated_quantity: f32) -> Result<()> {
     conn.execute("UPDATE inventory SET quantity = ?1 WHERE id = ?2",
     params![updated_quantity, item_id],
     )?;
@@ -334,7 +334,7 @@ pub fn update_inventory_quantity(conn: &Connection, item_id: f32, updated_quanti
 }
 
 // Function to update inventory cost_per_unit
-pub fn update_inventory_cost(conn: &Connection, item_id: f32, updated_cost: f32) -> Result<()> {
+pub fn update_inventory_cost(conn: &Connection, item_id: i32, updated_cost: f32) -> Result<()> {
     conn.execute("UPDATE inventory SET cost_per_unit = ?1 WHERE id = ?2",
     params![updated_cost, item_id],
     )?;
